Add API to easily obtain OS temporary file location
Also:
* Add $api.fp.curry
* Separate wo filesystem APIs into separate file
* Improve README.md, README.fifty.ts
* Rename $api.fp.mapping to $api.fp.Mapping
* Improve $api.fp documentation
