+++ README.fifty.ts
@@ -77,6 +77,11 @@
  *
  * ## Using SLIME
  *
+ * ### APIs for all platforms
+ *
+ * SLIME provides the {@link slime.$api.Global | `$api`} object to all code loaded with SLIME, which provides a number of
+ * general-purpose constructs, including a functional programming module available as {@link slime.$api.fp.Exports | $api.fp}.
+ *
  * ### SLIME definitions (documentation and testing)
  *
  * SLIME has the concept of a _definition_, which is a construct that provides both documentation and a test suite for a particular
