+++ README.md
@@ -13,9 +13,9 @@ The SLIME project provides tools for JavaScript development on several platforms
 * Writing Java-enabled JavaScript scripts via the `jsh` scripting environment
 * Writing JavaScript servlets for Java servlet containers (Tomcat, Jetty, etc.)
 * Writing front-end code: Chrome, Firefox, Safari
-* Using JXA, the macOS JavaScript automation framework
+* (requires repair) ~~Using JXA, the macOS JavaScript automation framework~~
 
-The Java-based environments support Mozilla [Rhino](https://github.com/mozilla/rhino) and OpenJDK [Nashorn](https://github.com/openjdk/nashorn). [GraalJS](https://github.com/oracle/graaljs) support is [under development](https://github.com/davidpcaldwell/slime/projects/10).
+The Java-based environments support Mozilla [Rhino](https://github.com/mozilla/rhino) and OpenJDK [Nashorn](https://github.com/openjdk/nashorn), including standalone Nashorn for JDK 15 and up. [GraalJS](https://github.com/oracle/graaljs) support is [under development](https://github.com/davidpcaldwell/slime/projects/10).
 
 ## Getting started: run a `jsh` script without installing anything
